split editor
